+++ curve_editor/dll_export.cpp
@@ -1,7 +1,11 @@
 #include <format>
 #include <aviutl.hpp>
 #include "constants.hpp"
-#include "func_filter.hpp"
+#include "filter_init.hpp"
+#include "filter_exit.hpp"
+#include "filter_project_load.hpp"
+#include "filter_project_save.hpp"
+#include "filter_wndproc.hpp"
