+++ curve_editor/filter_exit.cpp
@@ -1,0 +1,15 @@
+#include "filter_exit.hpp"
+#include <mkaul/include/graphics.hpp>
+#include "config.hpp"
+
+
+
+namespace cved {
+	BOOL filter_exit(AviUtl::FilterPlugin* fp) {
+		// 設定をJSONに保存
+		global::config.save_json();
+		mkaul::graphics::Factory::shutdown();
+
+		return TRUE;
+	}
+}
