+++ curve_editor/filter_exit.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <aviutl/FilterPlugin.hpp>
+
+
+
+namespace cved {
+	BOOL filter_exit(AviUtl::FilterPlugin* fp);
+}
