+++ curve_editor/filter_init.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <aviutl/FilterPlugin.hpp>
+
+
+
+namespace cved {
+	BOOL filter_init(AviUtl::FilterPlugin* fp);
+}
