+++ curve_editor/filter_project_load.cpp
@@ -1,0 +1,9 @@
+#include "filter_project_load.hpp"
+
+
+
+namespace cved {
+	BOOL filter_project_load(AviUtl::FilterPlugin* fp, AviUtl::EditHandle* editp, void* data, int32_t size) {
+		return TRUE;
+	}
+}
