+++ curve_editor/filter_project_save.cpp
@@ -1,0 +1,9 @@
+#include "filter_project_save.hpp"
+
+
+
+namespace cved {
+	BOOL filter_project_save(AviUtl::FilterPlugin* fp, AviUtl::EditHandle* editp, void* data, int32_t* size) {
+		return TRUE;
+	}
+}
