+++ curve_editor/filter_wndproc.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <aviutl/FilterPlugin.hpp>
+
+
+
+namespace cved {
+	BOOL filter_wndproc(HWND hwnd, UINT message, WPARAM wparam, LPARAM lparam, AviUtl::EditHandle* editp, AviUtl::FilterPlugin* fp);
+}
